sever dependency on synth dataset fixtures
the open_zarr mapper is very simple, so this test is unlikely to break
now, if our data definitions change.
