+++ external/fv3fit/tests/training/conftest.py
@@ -1,17 +1,11 @@
 from typing import Iterable, Sequence, Optional
-from synth import (  # noqa: F401
-    dataset_fixtures_dir,
-    data_source_name,
-    nudging_dataset_path,
-    fine_res_dataset_path,
-    data_source_path,
-    grid_dataset,
-)
 import xarray as xr
 from fv3fit._shared import ModelTrainingConfig, load_data_sequence
 import pytest
-import tempfile
-import yaml
+
+import numpy as np
+import xarray
+import cftime
 
 
 @pytest.fixture(params=[None])
@@ -29,38 +23,49 @@ def output_variables() -> Iterable[str]:
     return ["dQ1", "dQ2"]
 
 
-def get_batch_kwargs(data_source_name: str) -> dict:  # noqa: F811
-    if data_source_name == "nudging_tendencies":
-        return {
-            "needs_grid": False,
-            "res": "c8_random_values",
-            "timesteps_per_batch": 1,
-            "mapping_function": "open_merged_nudged",
-            "timesteps": ["20160801.001500"],
-            "mapping_kwargs": {
-                "i_start": 0,
-                "rename_vars": {
-                    "air_temperature_tendency_due_to_nudging": "dQ1",
-                    "specific_humidity_tendency_due_to_nudging": "dQ2",
-                },
-            },
-        }
-    elif data_source_name == "fine_res_apparent_sources":
-        return {
-            "needs_grid": False,
-            "res": "c8_random_values",
-            "timesteps_per_batch": 1,
-            "mapping_function": "open_fine_res_apparent_sources",
-            "timesteps": ["20160801.001500"],
-            "mapping_kwargs": {
-                "rename_vars": {
-                    "delp": "pressure_thickness_of_atmospheric_layer",
-                    "grid_xt": "x",
-                    "grid_yt": "y",
-                    "pfull": "z",
-                }
-            },
-        }
+@pytest.fixture
+def data_info(tmpdir):
+
+    # size needs to be 48 or an error happens. Is there a hardcode in fv3fit
+    # someplace?...maybe where the grid data is loaded?
+    x, y, z, tile, time = (8, 8, 79, 6, 2)
+    arr = np.zeros((time, tile, z, y, x))
+    arr_surf = np.zeros((time, tile, y, x))
+    dims = ["time", "tile", "z", "y", "x"]
+    dims_surf = ["time", "tile", "y", "x"]
+
+    data = xarray.Dataset(
+        {
+            "specific_humidity": (dims, arr),
+            "air_temperature": (dims, arr),
+            "downward_shortwave": (dims_surf, arr_surf),
+            "net_shortwave": (dims_surf, arr_surf),
+            "downward_longwave": (dims_surf, arr_surf),
+            "dQ1": (dims, arr),
+            "dQ2": (dims, arr),
+            "dQu": (dims, arr),
+            "dQv": (dims, arr),
+        },
+        coords={
+            "time": [
+                cftime.DatetimeJulian(2016, 8, 1),
+                cftime.DatetimeJulian(2016, 8, 2),
+            ]
+        },
+    )
+
+    data.to_zarr(str(tmpdir), consolidated=True)
+    return dict(
+        data_path=str(tmpdir),
+        batch_kwargs=dict(
+            mapping_function="open_zarr",
+            timesteps=["20160801.000000"],
+            needs_grid=False,
+            res="c8_random_values",
+            timesteps_per_batch=1,
+        ),
+        validation_timesteps=["20160802.000000"],
+    )
 
 
 @pytest.fixture
@@ -69,7 +74,7 @@ def train_config(
     hyperparameters: dict,
     input_variables: Iterable[str],
     output_variables: Iterable[str],
-    data_source_name,
+    data_info,
     validation_timesteps: Optional[Sequence[str]],
 ) -> ModelTrainingConfig:
     return ModelTrainingConfig(
@@ -79,7 +84,7 @@ def train_config(
         input_variables=input_variables,
         output_variables=output_variables,
         batch_function="batches_from_geodata",
-        batch_kwargs=get_batch_kwargs(data_source_name),
+        batch_kwargs=data_info["batch_kwargs"],
         scaler_type="standard",
         scaler_kwargs={},
         additional_variables=None,
@@ -90,14 +95,7 @@ def train_config(
 
 @pytest.fixture
 def training_batches(
-    data_source_path: str, train_config: ModelTrainingConfig,  # noqa: F811
+    data_info: str, train_config: ModelTrainingConfig,  # noqa: F811
 ) -> Sequence[xr.Dataset]:
-    batched_data = load_data_sequence(data_source_path, train_config)
+    batched_data = load_data_sequence(data_info["data_path"], train_config)
     return batched_data
-
-
-@pytest.fixture
-def data_info(data_source_path, data_source_name):
-    return dict(
-        data_path=data_source_path, batch_kwargs=get_batch_kwargs(data_source_name)
-    )
